Fix that failed to get the fontSize from gsettings

+++ src/processlistwidget.h
@@ -135,7 +135,7 @@ private:
     QPixmap m_downArrowPixmap;
     QPixmap m_upArrowPixmap;
 
-    int fontSize;
+    float fontSize;
 
     QGSettings *qtSettings;
     QGSettings *fontSettings;
